Modelling/Deploy/emergency/requirements.txt
Flask
dill
numpy
scikit-learn
google-cloud-bigquery==3.11.0
db-dtypes
gunicorn==20.1.0